requirements.txt
asgiref>=3.2.3
attrs>=19.3.0
backcall>=0.1.0
bleach>=3.1.0
blis>=0.4.1
certifi>=2019.9.11
chardet>=3.0.4
colorama>=0.4.1
cycler>=0.10.0
cymem>=2.0.2
decorator>=4.4.1
defusedxml>=0.6.0
dj-database-url>=0.5.0
Django>=3.0.1
django-heroku>=0.3.1
entrypoints>=0.3
gunicorn>=20.0.4
idna>=2.8
importlib-metadata>=0.23
ipykernel>=5.1.3
ipython>=7.9.0
ipython-genutils>=0.2.0
ipywidgets>=7.5.1
jedi>=0.15.1
Jinja2>=2.10.3
joblib>=0.14.0
jsonschema>=3.1.1
jupyter>=1.0.0
jupyter-client>=5.3.4
jupyter-console>=6.0.0
jupyter-core>=4.6.1
kiwisolver>=1.1.0
lxml>=4.4.2
MarkupSafe>=1.1.1
matplotlib>=3.1.2
mistune>=0.8.4
more-itertools>=7.2.0
murmurhash>=1.0.2
mysql-connector-python>=8.0.18
nbconvert>=5.6.1
nbformat>=4.4.0
notebook>=6.0.2
numpy>=1.17.3
pandas>=0.25.3
pandocfilters>=1.4.2
parso>=0.5.1
pickleshare>=0.7.5
Pillow>=6.2.1
plac>=1.1.3
preshed>=3.0.2
prometheus-client>=0.7.1
prompt-toolkit>=2.0.10
psycopg2>=2.8.4
Pygments>=2.4.2
pyparsing>=2.4.6
pyrsistent>=0.15.5
python-dateutil>=2.8.1
pytz>=2019.3
pywinpty>=0.5.5
pyzmq>=18.1.0
qtconsole>=4.5.5
requests>=2.22.0
scikit-learn>=0.21.3
scipy>=1.3.1
seaborn>=0.9.0
Send2Trash>=1.5.0
six>=1.13.0
sklearn>=0.0
spacy>=2.2.2
sqlparse>=0.3.0
srsly>=0.2.0
terminado>=0.8.2
testpath>=0.4.4
thinc>=7.3.1
tornado>=6.0.3
tqdm>=4.37.0
traitlets>=4.3.3
urllib3>=1.25.6
virtualenv>=16.7.9
wasabi>=0.4.0
wcwidth>=0.1.7
webencodings>=0.5.1
whitenoise>=5.0.1
whois>=0.9.5
widgetsnbextension>=3.5.1
xlrd>=1.2.0
youtube-transcript-api>=0.2.1
zipp>=0.6.0
